tests: split off connect_network fn for reuse
